Updating SecurityProperties with new Role And Group models
