+++ src/main/java/org/akhq/configs/Group.java
@@ -3,11 +3,10 @@ package org.akhq.configs;
 import lombok.Data;
 
 import java.util.List;
-import java.util.Map;
 
 @Data
 public class Group {
-    String name;
-    List<String> roles;
-    Map<String, List<String>> attributes;
+    String role;
+    List<String> patterns = List.of(".*");
+    List<String> clusters = List.of(".*");
 }
